move prismaClient into services folder
